Polished soapclient.
Added plenty of explanation to the file.
Also cleaned up where which escaping happens:

* The crm importer makes sure html is unescape so we have proper \n etc.
* the latex jinja environment has now filters for \n -> \\ etc
* the flask jinja template for main.html now substution \n -> <br />

This way the the new lines etc. are represented correctly within the app
and rendered properly for latex and html

Also added some simple tests for the soapclient.
(Very basic, since data can not be veryfied it just checks if it doesnt crash)

To run the tests
```
pip install nose2
nose2
```